requirements.txt
git+https://github.com/Force1ess/python-pptx
jinja2
openai
google-generativeai
tenacity
transformers
FlagEmbedding
peft
timm
einops
vllm
pillow
googlesearch-python
socksio
pdf2image
numpy
rich
sentencepiece
lxml
torch
torchvision
pytorch_fid
# binary: poppler-utils libreoffice
